position changes of nodes refactored completely.
whenever a node is being dragged the current location is synced
into the state store @ movingNode. this is done in order to be
able to show the current location to connections at all times.
when the dragging stops, the last position is committed to
userlog. this enables timetravel for node pos changes.